refactor(e2e): use folder index entry points for all split utils
Remove redundant top-level common.ts and ui-helper.ts facades so split
modules match api-helper and kube-client (folder + index only). Add
ui-helper/index.ts and move msgraph-helper split into msgraph-helper/.
