--- e2e-tests/playwright/utils/common.ts
@@ -1,1 +1,0 @@
-export { Common, setupBrowser, teardownBrowser } from "./common/index";
