--- e2e-tests/playwright/utils/ui-helper.ts
@@ -1,1 +1,0 @@
-export { UIhelper } from "./ui-helper/class";
