+++ e2e-tests/playwright/utils/ui-helper/index.ts
@@ -1,0 +1,1 @@
+export { UIhelper } from "./class";
